chore: moved content to homepagw
