+++ src/app/page.js
@@ -1,5 +1,21 @@
 'use client';
 
-export default function Home() {
-  return <div className='min-h-screen'></div>;
-}
+import { Hero } from '@/components/HeroSection/Hero';
+import { AboutUsMain } from '@/components/AboutUs/Main';
+import { EventSectionMain } from '@/components/EventsSection/wrapperComponents/Main';
+import { StatisticsMain } from '@/components/Statistics/wrapper/Main';
+import { SponsorSection } from '@/components/Sponsors/sponsors';
+
+const Page = () => {
+  return (
+    <>
+      <Hero />
+      <AboutUsMain />
+      <StatisticsMain />
+      <EventSectionMain />
+      <SponsorSection />
+    </>
+  );
+};
+
+export default Page;
