+++ src/app/playground/page.js
@@ -1,10 +1,10 @@
 'use client';
 
-import Hero from '@/components/Hero Section/Hero';
+import { Hero } from '@/components/HeroSection/Hero';
 import { AboutUsMain } from '@/components/AboutUs/Main';
 import { EventSectionMain } from '@/components/EventsSection/wrapperComponents/Main';
 import { StatisticsMain } from '@/components/Statistics/wrapper/Main';
-import SponsorSection from '@/components/Sponsers/sponsors';
+import { SponsorSection } from '@/components/Sponsors/sponsors';
 
 const Page = () => {
   return (
