+++ src/components/HeroSection/styles.jsx
